DVD_RentalApp/api/serializers.py
# from django.contrib.auth import get_user_model
from rest_framework import serializers
from movies.models import Movie

class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = '__all__'

