requirements.txt
asgiref==3.5.2
attrs==22.1.0
coverage==3.6
Django==4.1
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
